Unified ItemContentDetails with iOS and macOS
