--- Shared/View/Details/macOS/ItemContentDetailsView.swift
@@ -1,367 +1,0 @@
-//
-//  ItemContentDetailsView.swift
-//  CronicaMac
-//
-//  Created by Alexandre Madeira on 02/11/22.
-//
-
-import SwiftUI
-import SDWebImageSwiftUI
-
-#if os(macOS)
-struct ItemContentDetailsView: View {
-    let id: Int
-    let title: String
-    let type: MediaType
-    @StateObject private var viewModel: ItemContentViewModel
-    @State private var showConfirmation = false
-    @State private var showSeasonConfirmation = false
-    private var itemUrl: URL
-    @State private var actionMessageConfirmation = ""
-    @State private var actionImageConfirmation = ""
-    @State private var showActionConfirmation = false
-    var handleToolbarOnPopup = false
-    @State private var showCustomList = false
-    init(id: Int, title: String, type: MediaType, handleToolbarOnPopup: Bool = false) {
-        self.id = id
-        self.title = title
-        self.type = type
-        _viewModel = StateObject(wrappedValue: ItemContentViewModel(id: id, type: type))
-        self.itemUrl = URL(string: "https://www.themoviedb.org/\(type.rawValue)/\(id)")!
-        self.handleToolbarOnPopup = handleToolbarOnPopup
-        
-    }
-    var body: some View {
-        ZStack {
-            if viewModel.isLoading { ProgressView() }
-            VStack {
-                ScrollView {
-                    headerView
-                    
-                    TrailerListView(trailers: viewModel.content?.itemTrailers)
-                    
-                    if let seasons = viewModel.content?.itemSeasons {
-                        SeasonList(showID: id, numberOfSeasons: seasons)
-                            .padding(.zero)
-                    }
-                    
-                    WatchProvidersList(id: id, type: type)
-                    
-                    CastListView(credits: viewModel.credits)
-                    
-                    ItemContentListView(items: viewModel.recommendations,
-                                        title: "Recommendations",
-                                        subtitle: "You may like",
-                                        addedItemConfirmation: $showConfirmation,
-                                        displayAsCard: true)
-                    
-                    InformationSectionView(item: viewModel.content)
-                        .padding()
-                    
-                    AttributionView()
-                }
-            }
-            .task {
-                await viewModel.load()
-            }
-            .navigationDestination(for: ItemContent.self) { item in
-                ItemContentDetailsView(id: item.id, title: item.itemTitle, type: item.itemContentMedia)
-            }
-            .navigationDestination(for: Person.self) { item in
-                PersonDetailsView(title: item.name, id: item.id)
-            }
-            .toolbar {
-                if handleToolbarOnPopup {
-                    ToolbarItem(placement: .status) {
-                        ViewThatFits {
-                            HStack {
-                                notificationButton
-                                watchButton
-                                favoriteButton
-                                shareButton
-                            }
-                            shareButton
-                        }
-                        
-                    }
-                } else {
-                    ToolbarItem {
-                        ViewThatFits {
-                            HStack {
-                                notificationButton
-                                watchButton
-                                favoriteButton
-                                if viewModel.isInWatchlist {
-                                    addToCustomListButton
-                                        .sheet(isPresented: $showCustomList) {
-                                            ItemContentCustomListSelector(item: $viewModel.watchlistItem, showView: $showCustomList)
-                                                .presentationDetents([.medium])
-                                                .frame(width: 500, height: 600, alignment: .center)
-                                        }
-                                }
-                                shareButton
-                            }
-                            shareButton
-                        }
-                        
-                    }
-                }
-            }
-            .navigationTitle(title)
-            ConfirmationDialogView(showConfirmation: $showConfirmation, message: "markedAsWatched")
-            ConfirmationDialogView(showConfirmation: $showSeasonConfirmation,
-                                   message: "Season Marked as Watched")
-            ConfirmationDialogView(showConfirmation: $showActionConfirmation,
-                                   message: actionMessageConfirmation,
-                                   image: actionImageConfirmation)
-        }
-        .background {
-            TranslucentBackground(image: viewModel.content?.cardImageLarge)
-        }
-    }
-    
-    private var watchButton: some View {
-        Button {
-            if viewModel.isWatched {
-                actionMessageConfirmation = "removedFromWatched"
-                actionImageConfirmation = "minus.circle"
-            } else {
-                actionMessageConfirmation = "markedAsWatched"
-                actionImageConfirmation = "checkmark.circle"
-            }
-            viewModel.update(.watched)
-            showActionConfirmation.toggle()
-            DispatchQueue.main.asyncAfter(deadline: .now() + 1.4) {
-                withAnimation {
-                    showActionConfirmation = false
-                    actionMessageConfirmation = ""
-                    actionImageConfirmation = ""
-                }
-            }
-        } label: {
-            Label(viewModel.isWatched ? "Remove from Watched" : "Mark as Watched",
-                  systemImage: viewModel.isWatched ? "minus.circle" : "checkmark.circle")
-        }
-        .keyboardShortcut("w", modifiers: [.option])
-        .disabled(viewModel.isLoading ? true : false)
-    }
-    
-    private var favoriteButton: some View {
-        Button {
-            if viewModel.isFavorite {
-                actionMessageConfirmation = "removedFromFavorites"
-                actionImageConfirmation = "heart.circle.fill"
-            } else {
-                actionMessageConfirmation = "markedAsFavorite"
-                actionImageConfirmation = "heart.circle"
-            }
-            viewModel.update(.favorite)
-            showActionConfirmation.toggle()
-            DispatchQueue.main.asyncAfter(deadline: .now() + 1.4) {
-                withAnimation {
-                    showActionConfirmation = false
-                    actionMessageConfirmation = ""
-                    actionImageConfirmation = ""
-                }
-            }
-        } label: {
-            Label(viewModel.isFavorite ? "Remove from Favorites" : "Mark as Favorite",
-                  systemImage: viewModel.isFavorite ? "heart.circle.fill" : "heart.circle")
-        }
-        .keyboardShortcut("f", modifiers: [.option])
-        .disabled(viewModel.isLoading ? true : false)
-    }
-    
-    private var shareButton: some View {
-        ShareLink(item: itemUrl)
-    }
-    
-    private var notificationButton: some View {
-        Button {
-            
-        } label: {
-            Image(systemName: viewModel.hasNotificationScheduled ? "bell.fill" : "bell")
-                .opacity(viewModel.isNotificationAvailable ? 1 : 0)
-                .accessibilityHidden(true)
-                .accessibilityLabel(viewModel.hasNotificationScheduled ? "Notification scheduled." : "No notification scheduled.")
-        }
-    }
-    
-    private var addToCustomListButton: some View {
-        Button {
-            if viewModel.watchlistItem == nil {
-                viewModel.fetchSavedItem()
-            }
-            showCustomList.toggle()
-        } label: {
-            Label("addToCustomList", systemImage: "rectangle.on.rectangle.angled")
-        }
-    }
-    
-    
-}
-#endif
-
-struct GlanceInfo: View {
-    var info: String?
-    var body: some View {
-        if let info {
-            Text(info)
-                .font(.callout)
-        }
-    }
-}
-
-
-struct LargerHeader: View {
-    let title: String
-    let type: MediaType
-    @EnvironmentObject var viewModel: ItemContentViewModel
-    var body: some View {
-#if os(macOS)
-        image
-#else
-        image
-            .clipShape(RoundedRectangle(cornerRadius: 20, style: .continuous))
-            .padding()
-            .shadow(radius: 5)
-#endif
-    }
-    
-    private var image: some View {
-        WebImage(url: viewModel.content?.cardImageOriginal)
-            .resizable()
-            .placeholder {
-                ZStack {
-                    Rectangle().fill(Color.gray.gradient)
-                    Image(systemName: type == .tvShow ? "tv" : "film")
-                        .foregroundColor(.secondary)
-                        .font(.title)
-                        .fontWeight(.semibold)
-                }
-                .frame(height: 500)
-                .padding(.zero)
-            }
-            .aspectRatio(contentMode: .fill)
-            .overlay {
-                ZStack {
-                    if viewModel.content?.cardImageOriginal != nil {
-                        VStack {
-                            Spacer()
-                            ZStack(alignment: .bottom) {
-                                Color.black.opacity(0.8)
-                                    .frame(height: 150)
-                                    .mask {
-                                        LinearGradient(colors: [Color.black,
-                                                                Color.black.opacity(0.924),
-                                                                Color.black.opacity(0.707),
-                                                                Color.black.opacity(0.383),
-                                                                Color.black.opacity(0)],
-                                                       startPoint: .bottom,
-                                                       endPoint: .top)
-                                    }
-                                Rectangle()
-                                    .fill(.ultraThinMaterial)
-                                    .frame(height: 140)
-                                    .mask {
-                                        LinearGradient(colors: [Color.black,
-                                                                Color.black.opacity(0.924),
-                                                                Color.black.opacity(0.707),
-                                                                Color.black.opacity(0.383),
-                                                                Color.black.opacity(0)],
-                                                       startPoint: .bottom,
-                                                       endPoint: .top)
-                                        .frame(height: 150)
-                                    }
-                            }
-                        }
-                    }
-                    VStack(alignment: .leading) {
-                        Spacer()
-#if os(macOS)
-                        informations
-                            .padding()
-#else
-                        ViewThatFits {
-                            informations.padding()
-                            compactInfo.padding(.bottom)
-                        }
-#endif
-                    }
-                }
-            }
-            .transition(.scale)
-    }
-    
-    private var informations: some View {
-        HStack(alignment: .bottom) {
-            VStack {
-                Text(title)
-                    .lineLimit(1)
-                    .font(.title)
-                    .foregroundColor(.white)
-                GlanceInfo(info: viewModel.content?.itemInfo)
-                    .padding(.bottom, 6)
-                    .foregroundColor(.white.opacity(0.8))
-                DetailWatchlistButton()
-                    .environmentObject(viewModel)
-            }
-            .frame(maxWidth: 600)
-            .padding(.horizontal)
-            Spacer()
-#if os(iOS)
-            OverviewBoxView(overview: viewModel.content?.itemOverview,
-                            title: "About",
-                            type: type)
-            .groupBoxStyle(TransparentGroupBox())
-            .padding([.horizontal, .top])
-            .frame(maxWidth: 500)
-#else
-            OverviewBoxView(overview: viewModel.content?.itemOverview,
-                            title: "About",
-                            type: .movie)
-            .foregroundColor(.white)
-            .padding([.horizontal, .top])
-            .frame(maxWidth: 500)
-#endif
-            Spacer()
-        }
-    }
-    
-    private var compactInfo: some View {
-        CenterHorizontalView {
-            VStack {
-                GlanceInfo(info: viewModel.content?.itemInfo)
-                    .padding(.bottom, 6)
-                    .foregroundColor(.white.opacity(0.8))
-                DetailWatchlistButton()
-                    .environmentObject(viewModel)
-            }
-        }
-    }
-}
-
-
-struct TransparentGroupBox: GroupBoxStyle {
-    func makeBody(configuration: Configuration) -> some View {
-        VStack {
-            HStack {
-                configuration.label
-                    .font(.headline)
-                    .foregroundColor(.white)
-                Spacer()
-            }
-            
-            configuration.content
-                .foregroundColor(.white)
-        }
-        .padding()
-        .background {
-            ZStack {
-                Rectangle().fill(.black.opacity(0.2))
-                Rectangle().fill(.ultraThinMaterial)
-            }
-            .clipShape(RoundedRectangle(cornerRadius: 16, style: .continuous))
-        }
-    }
-}
